requirements.txt
boto3==1.42.34
cryptography==46.0.5
fastapi==0.133.0
mysql-connector-python==9.4.0
pydantic==2.12.5
PyJWT==2.11.0
python-dotenv==1.2.1
starlette==0.52.1
python-multipart==0.0.20
requests==2.32.5
uvicorn==0.38.0
itsdangerous